Move FrozenSet to node_classes so it is accessible
importing astroid.objects causes curcular imports with manager
