+++ astroid/tests/unittest_inference.py
@@ -26,7 +26,6 @@ from astroid import arguments
 from astroid import decorators as decoratorsmod
 from astroid import exceptions
 from astroid import helpers
-from astroid import objects
 from astroid import test_utils
 from astroid import util
 from astroid.tests import resources
@@ -89,7 +88,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
     assertInferTuple = partialmethod(_assertInferElts, nodes.Tuple)
     assertInferList = partialmethod(_assertInferElts, nodes.List)
     assertInferSet = partialmethod(_assertInferElts, nodes.Set)
-    assertInferFrozenSet = partialmethod(_assertInferElts, objects.FrozenSet)
+    assertInferFrozenSet = partialmethod(_assertInferElts, nodes.FrozenSet)
 
     CODE = '''
         class C(object):
